// 6/task_6/src/mocks/types.ts
export type ItemType = {
    id: number,
    name: string,
    layout: number[][], // матрица, описывающая форму предмета, 1 - клетка предмета, 0 - пустая клетка, поддерживаются только прямоугольные формы предметов
    position: [number, number], // [row, col] - координаты ячейки, в которой рисуется предмет, отсчет клеток идет направо и вниз 
    rarity: number, // коэффициент редкости предмета, от 1 до 5 в порядке увеличения редкости
    image: string // путь к изображению предмета
}